+++ src/render.rs
@@ -141,6 +141,8 @@ impl<'a, W: std::io::Write> HeaderContext<'a, W> {
         let path = obj.get("path").and_then(JsonValue::as_str);
         if let Some(path) = path {
             cookie.set_path(path);
+        } else {
+            cookie.set_path("/");
         }
         let domain = obj.get("domain").and_then(JsonValue::as_str);
         if let Some(domain) = domain {
@@ -276,9 +278,9 @@ pub struct RenderContext<W: std::io::Write> {
     current_statement: usize,
 }
 
-const DEFAULT_COMPONENT: &str = "debug";
+const DEFAULT_COMPONENT: &str = "table";
 const PAGE_SHELL_COMPONENT: &str = "shell";
-const FRAGMENT_SHELL_COMPONENT: &str = "fragment-shell";
+const FRAGMENT_SHELL_COMPONENT: &str = "shell-empty";
 const DYNAMIC_COMPONENT: &str = "dynamic";
 const MAX_RECURSION_DEPTH: usize = 256;
 
@@ -287,43 +289,47 @@ impl<W: std::io::Write> RenderContext<W> {
         app_state: Arc<AppState>,
         layout_context: &LayoutContext,
         mut writer: W,
-        mut initial_row: JsonValue,
+        initial_row: JsonValue,
     ) -> anyhow::Result<RenderContext<W>> {
         log::debug!("Creating the shell component for the page");
-        let shell_component = if layout_context.is_embedded {
-            FRAGMENT_SHELL_COMPONENT
-        } else {
-            PAGE_SHELL_COMPONENT
-        };
-        let mut shell_renderer = Self::create_renderer(shell_component, Arc::clone(&app_state), 0)
-            .await
-            .with_context(|| "The shell component should always exist")?;
 
-        let mut initial_component =
-            Some(get_object_str(&initial_row, "component").unwrap_or(DEFAULT_COMPONENT));
-        let mut shell_properties = JsonValue::Null;
-        match initial_component {
-            Some(PAGE_SHELL_COMPONENT) => {
-                shell_properties = initial_row.take();
-                initial_component = None;
-            },
-            Some(DYNAMIC_COMPONENT) => {
-                let dynamic_properties = Self::extract_dynamic_properties(&initial_row)?;
-                for prop in dynamic_properties {
-                    match get_object_str(&prop, "component") {
-                        None | Some(PAGE_SHELL_COMPONENT) => {
-                            shell_properties = prop.into_owned();
-                            initial_component = None;
-                        },
-                        _ => bail!("Dynamic components at the top level are not supported, except for setting the shell component properties"),
-                    }
-                }
-            },
-            _ => log::trace!("The first row is not a shell component, so we will render a shell with default properties"),
+        let mut initial_rows =
+            if get_object_str(&initial_row, "component") == Some(DYNAMIC_COMPONENT) {
+                Self::extract_dynamic_properties(&initial_row)?
+            } else {
+                vec![Cow::Borrowed(&initial_row)]
+            };
+
+        if !initial_rows
+            .first()
+            .and_then(|c| get_object_str(c, "component"))
+            .is_some_and(Self::is_shell_component)
+        {
+            let default_shell = if layout_context.is_embedded {
+                FRAGMENT_SHELL_COMPONENT
+            } else {
+                PAGE_SHELL_COMPONENT
+            };
+            let added_row = json!({"component": default_shell});
+            log::trace!(
+                "No shell component found in the first row. Adding the default shell: {added_row}"
+            );
+            initial_rows.insert(0, Cow::Owned(added_row));
         }
+        let mut rows_iter = initial_rows.into_iter().map(Cow::into_owned);
 
-        log::debug!("Rendering the shell with properties: {shell_properties}");
-        shell_renderer.render_start(&mut writer, shell_properties)?;
+        let shell_row = rows_iter
+            .next()
+            .expect("shell row should exist at this point");
+        let mut shell_renderer = Self::create_renderer(
+            get_object_str(&shell_row, "component").expect("shell should exist"),
+            Arc::clone(&app_state),
+            0,
+        )
+        .await
+        .with_context(|| "The shell component should always exist")?;
+        log::debug!("Rendering the shell with properties: {shell_row}");
+        shell_renderer.render_start(&mut writer, shell_row)?;
 
         let mut initial_context = RenderContext {
             app_state,
@@ -334,14 +340,15 @@ impl<W: std::io::Write> RenderContext<W> {
             current_statement: 1,
         };
 
-        if let Some(component) = initial_component {
-            log::trace!("The page starts with a component without a shell: {component}");
-            initial_context
-                .open_component_with_data(component, &initial_row)
-                .await?;
+        for row in rows_iter {
+            initial_context.handle_row(&row).await?;
         }
 
         Ok(initial_context)
+    }
+
+    fn is_shell_component(component: &str) -> bool {
+        component.starts_with(PAGE_SHELL_COMPONENT)
     }
 
     async fn current_component(&mut self) -> anyhow::Result<&mut SplitTemplateRenderer> {
@@ -375,6 +382,11 @@ impl<W: std::io::Write> RenderContext<W> {
                 bail!("The {component_name} component cannot be used after data has already been sent to the client's browser. \
                 This component must be used before any other component. \
                 To fix this, either move the call to the '{component_name}' component to the top of the SQL file, or create a new SQL file where '{component_name}' is the first component.");
+            }
+            (_, Some(c)) if Self::is_shell_component(c) => {
+                bail!("There cannot be more than a single shell per page. \n\
+                You are trying to open the {c:?} component, but a shell component is already opened for the current page. \n\
+                You can fix this by removing the extra shell component, or by moving this component to the top of the SQL file, before any other component that displays data. \n")
             }
             (_current_component, Some(new_component)) => {
                 self.open_component_with_data(new_component, &data).await?;