+++ tests/sql_test_files/it_works_dynamic_toplevel.sql
@@ -1,0 +1,8 @@
+-- Checks that we can have a page with a single dynamic component containing multiple children
+select 'dynamic' as component,
+    '[
+        {"component":"text"},
+        {"contents":"Hello, ", "bold":true},
+        {"component":"text"},
+        {"contents":"It works !", "bold":true}
+    ]' as properties;