+++ tests/sql_test_files/it_works_without_a_shell.sql
@@ -1,0 +1,2 @@
+select 'shell-empty' as component;
+select 'text' as component, '<!DOCTYPE html><html><head><title>My page</title></head><body><h1>It works !</h1></body></html>' as html;